feat: ReviewItem을 사용한 HospitalReviewWrapper 작성
