+++ src/app/profile/_component/ProfileContent/ProfileContent.tsx
@@ -7,7 +7,7 @@ import { useRouter, useSearchParams } from "next/navigation";
 import { useGetMyComment, useGetMyPost } from "@api/domain/mypage/hook";
 import { PATH } from "@route/path";
 import { formatTime } from "@shared/util/formatTime";
-import HospitalReview from "@shared/component/HospitalReview";
+import HospitalReviewWrapper from "@shared/component/HospitalReviewWrapper";
 
 interface ProfileContentPropTypes {
   tab: ActiveTabType;
@@ -44,7 +44,7 @@ const ProfileContent = ({ tab }: ProfileContentPropTypes) => {
   const renderContent = (tab: ActiveTabType) => {
     switch (tab) {
       case "review":
-        return <HospitalReview />;
+        return <HospitalReviewWrapper />;
       case "post":
         if (!profilePosts?.length) {
           return <div className={styles.nothingContent}>{"아직 작성한 게시글이 없어요."}</div>;
